requirements.txt
tk~=0.1.0
future~=0.18.2
Pillow~=8.2.0
pyinstaller~=4.3